+++ include/ParallaxGenDirectory/ParallaxGenDirectory.hpp
@@ -10,34 +10,23 @@
 
 class ParallaxGenDirectory : public BethesdaDirectory {
 private:
-	// list of mesh block names to ignore (if the path has any of these as a component it is ignored)
-	static inline const std::vector<std::wstring> mesh_blocklist = {
-		L"actors",
-		L"cameras",
-		L"decals",
-		L"effects",
-		L"interface",
-		L"loadscreenart",
-        L"lod",
-		L"magic",
-		L"markers",
-		L"mps",
-		L"sky",
-		L"water"
-	};
+	std::filesystem::path EXE_PATH;
+	static inline const std::wstring LO_PGCONFIG_PATH = L"parallaxgen";
 
 	std::vector<std::filesystem::path> heightMaps;
 	std::vector<std::filesystem::path> complexMaterialMaps;
 	std::vector<std::filesystem::path> meshes;
+  std::vector<nlohmann::json> truePBRConfigs;
+  
+	nlohmann::json PG_config;
 
 	static inline const std::vector<std::string> truePBR_filename_fields = { "match_normal", "match_diffuse", "rename" };
-	std::vector<nlohmann::json> truePBRConfigs;
 
 public:
 	static inline const std::filesystem::path default_cubemap_path = "textures\\cubemaps\\dynamic1pxcubemap_black.dds";
 
 	// constructor - calls the BethesdaDirectory constructor
-	ParallaxGenDirectory(BethesdaGame bg);
+	ParallaxGenDirectory(BethesdaGame bg, std::filesystem::path EXE_PATH);
 
 	// searches for height maps in the data directory
 	void findHeightMaps();
@@ -47,6 +36,9 @@ public:
 	void findMeshes();
 	// find truepbr config files
 	void findTruePBRConfigs();
+
+	// get the parallax gen config
+	void loadPGConfig(bool load_default);
 
 	// add methods
 	void addHeightMap(std::filesystem::path path);
@@ -65,6 +57,11 @@ public:
 	const std::vector<std::filesystem::path> getComplexMaterialMaps() const;
 	const std::vector<std::filesystem::path> getMeshes() const;
 	const std::vector<nlohmann::json> getTruePBRConfigs() const;
+
+	// Helpers
+	static void merge_json_smart(nlohmann::json& target, const nlohmann::json& source);
+	static std::vector<std::wstring> jsonArrayToWString(const nlohmann::json& json_array);
+	static void replaceForwardSlashes(nlohmann::json& j);
 };
 
 #endif